// src/enums/hooks.enum.ts
export enum HooksEnum {
    onInit = 'onInit', 
    onDestroy = 'onDestroy', 
    onViewChange = 'onViewChange', 
    beforeViewInit = 'beforeViewInit', 
    afterViewInit = 'afterViewInit', 
    connectedCallback = 'connectedCallback', 
    disconnectedCallback = 'disconnectedCallback', 
    onPropsChange = 'onPropsChange', 
    attributeChangedCallback = 'attributeChangedCallback', 
    adoptedCallback = 'adoptedCallback',
    onChangeDetection = 'onChangeDetection'
}